+++ src/constants/api-endpoints.js
@@ -3,7 +3,6 @@ const indexPage = 'index_page'
 const indexPageCategories = 'index_page_categories'
 const posts = 'posts'
 const users = 'users'
-const bookmarks = 'bookmarks'
 
 export default {
   topics,
@@ -11,5 +10,4 @@ export default {
   indexPageCategories,
   posts,
   users,
-  bookmarks,
 }