+++ src/index.js
@@ -6,12 +6,10 @@ import indexPage from './reducers/index-page'
 import { denormalizePosts, denormalizeTopics } from './utils/denormalize-asset'
 import { post, posts } from './reducers/posts'
 import { topic, topics } from './reducers/topics'
-import bookmarks from './reducers/bookmarks'
 import { fetchIndexPageContent, fetchCategoriesPostsOnIndexPage } from './actions/index-page'
 import { fetchAFullPost, fetchListedPosts, fetchEditorPickedPosts,
   fetchInfographicPostsOnIndexPage, fetchPhotographyPostsOnIndexPage } from './actions/posts'
 import { fetchAFullTopic, fetchTopics, fetchTopicsOnIndexPage } from './actions/topics'
-import { createBookmark, getBookmarks, deleteBookmark, getCurrentBookmark } from './actions/bookmarks'
 
 export default {
   actions: {
@@ -25,10 +23,6 @@ export default {
     fetchTopicsOnIndexPage,
     fetchInfographicPostsOnIndexPage,
     fetchPhotographyPostsOnIndexPage,
-    createBookmark,
-    getBookmarks,
-    deleteBookmark,
-    getCurrentBookmark,
   },
   reducers: {
     entities,
@@ -37,7 +31,6 @@ export default {
     topic,
     topics,
     indexPage,
-    bookmarks,
   },
   reduxStateFields,
   utils: {